+++ sciris/sc_parallel.py
@@ -77,7 +77,7 @@ def loadbalancer(maxload=None, index=None, interval=None, maxtime=None, label=No
     return None
 
 
-def parallelize(func, iterarg=None, iterkwargs=None, args=None, kwargs=None, ncpus=None, maxload=None, interval=None, parallelizer=None, **func_kwargs):
+def parallelize(func, iterarg=None, iterkwargs=None, args=None, kwargs=None, ncpus=None, maxload=None, interval=None, parallelizer=None, serial=False, **func_kwargs):
     '''
     Main method for parallelizing a function.
 
@@ -106,6 +106,8 @@ def parallelize(func, iterarg=None, iterkwargs=None, args=None, kwargs=None, ncp
         maxload (float): maximum CPU load to use (not used if ncpus is specified)
         interval (float): number of seconds to pause between starting processes for checking load (not used if ncpus is specified)
         parallelizer (func): alternate parallelization function instead of multiprocess.Pool.map()
+        serial (bool): whether to skip parallelization run in serial (useful for debugging)
+        func_kwargs (dict): merged with kwargs (see above)
 
     Returns:
         List of outputs from each process
@@ -177,18 +179,14 @@ def parallelize(func, iterarg=None, iterkwargs=None, args=None, kwargs=None, ncp
             results = sc.parallelize(func=f, iterarg=[(1,2),(2,3),(3,4)])
             print(results)
 
-    Version: 2020nov30
+    New in version 1.1.1: "serial" argument.
     '''
     # Handle maxload
     if ncpus is None and maxload is None:
         ncpus = mp.cpu_count()
 
     # Handle kwargs
-    if func_kwargs:
-        if kwargs is None:
-            kwargs = func_kwargs
-        else:
-            kwargs.update(func_kwargs)
+    kwargs = ut.mergedicts(kwargs, func_kwargs)
 
     # Handle iterarg and iterkwargs
     niters = 0
@@ -251,7 +249,9 @@ def parallelize(func, iterarg=None, iterkwargs=None, args=None, kwargs=None, ncp
 
     # Run simply using map -- no advantage here to using Process/Queue
     try:
-        if parallelizer is None: # Standard usage: use the default map() function
+        if serial: # Run in serial
+            outputlist = list(map(_parallel_task, argslist))
+        elif parallelizer is None: # Standard usage: use the default map() function
             multipool = mp.Pool(processes=ncpus)
             outputlist = multipool.map(_parallel_task, argslist)
             multipool.close()