+++ tests/test_odict.py
@@ -4,6 +4,7 @@ Define a suite of tests for the odict
 
 import numpy as np
 import sciris as sc
+import pytest
 
 
 def printexamples(examples):
@@ -37,10 +38,12 @@ def test_main():
 def test_insert():
     sc.heading('Insert:')
     z = sc.odict()
-    z['foo'] = 1492
-    z.insert(1604)
-    z.insert(0, 'ganges', 1444)
-    z.insert(2, 'midway', 1234)
+    z['foo'] = 11
+    z.insert(1123)
+    z.insert('bar', 112358)
+    z.insert(0, 'cat', 11235813)
+    z.insert(2, 'dog', 1123581321)
+    assert len(z) == 5
     printexamples([z])
     return
 
@@ -113,7 +116,9 @@ def test_repr():
     for i in range(n_entries):
         key = f'key{i:03d}'
         qq[key] = i**2
-    print(qq)
+    qq.disp(maxitems=15)
+    qq[:] = 4
+    assert all([v==4 for v in qq.values()])
     return
 
 
@@ -121,16 +126,21 @@ def test_default():
     sc.heading('Testing default')
 
     o = sc.odict(foo=[1,2,3], defaultdict=list)
+    p = sc.objdict(foo=[1,2,3], defaultdict=list)
     o['bar'].extend([1,2,3])
-    assert o[0] == o[1]
+    p.bar.extend([1,2,3])
+    assert o[0] == o[1] == p.foo == p.bar
 
     lam = lambda: 44
     o = sc.odict(defaultdict=lam)
     assert o['anything'] == lam()
 
-    # Testing objdict
+    # Testing nesting
     j = sc.objdict(a=1, b=2, defaultdict='nested')
-    j.c.d.e = 3
+    j.c.d.e = 12345
+    k = sc.odict(a=1, b=2, defaultdict='nested')
+    k['c']['d']['e'] = 12345
+    assert j['c']['d']['e'] == k['c']['d']['e']
     print(j)
 
     return
@@ -189,10 +199,49 @@ def test_other():
 
 def test_asobj():
     sc.heading('Testing objdict/asobj')
+
+    # Testing standard usage
     d = dict(foo=1, bar=2)
     d_obj = sc.asobj(d)
     d_obj.foo = 10
     assert d_obj.bar == 2
+
+    print('Testing derived classes')
+    class MyObjDict(sc.objdict):
+        def __init__(self, x):
+            self.x = x
+        def square(self):
+            return self.x**2
+        def __repr__(self):
+            output = sc.objrepr(self)
+            output += sc.objdict.__repr__(self)
+            return output
+
+    class MyObjObj(type(d_obj)):
+        def __init__(self, x):
+            self.x = x
+        def square(self):
+            return self.x**2
+
+    myobjdict = MyObjDict(3)
+    myobjobj = MyObjObj(4)
+    assert myobjdict.x == 3
+    assert myobjobj.x == 4
+    assert myobjdict.square() == 9
+    assert myobjobj.square() == 16
+    print(myobjdict)
+    print(myobjobj)
+    repr(myobjdict)
+    del myobjdict.x
+    myobjdict.setattribute('x', 10)
+    assert myobjdict.x == 10
+    assert getattr(myobjdict, 'x') == myobjdict.getattribute('x') ==  10
+
+    with pytest.raises(ValueError):
+        myobjdict.x = 'cannot change actual attribute'
+    with pytest.raises(AttributeError):
+        myobjdict.setattribute('keys', 4)
+
     return